# docs/swagger.yaml
basePath: /api/v1
definitions:
  http.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  http.SubscriptionResponse:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      subscription_id:
        type: string
      user_id:
        type: string
    type: object
  http.TotalCostResponse:
    properties:
      total_cost:
        type: integer
    type: object
  http.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    type: object
info:
  contact: {}
  description: Subscriptions CRUDL Service.
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Returns a list of subscriptions with the optional filtering
      parameters:
      - description: Filter by user id
        in: query
        name: user_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      - description: Period start, MM-YYYY
        in: query
        name: period_from
        type: string
      - description: Period end, MM-YYYY
        in: query
        name: period_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.SubscriptionResponse'
            type: array
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Creates a new subscription record
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/http.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Deletes a subscription record
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      description: Returns a subscription record by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.SubscriptionResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a subscription by id
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Updates a subscription record
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/http.UpdateSubscriptionRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a subscription
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Counts a total cost of subscriptions with the filtering
      parameters:
      - description: Filter by user id
        in: query
        name: user_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      - description: Period start, MM-YYYY
        in: query
        name: period_from
        type: string
      - description: Period end, MM-YYYY
        in: query
        name: period_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TotalCostResponse'
      summary: Total cost of subscriptions
      tags:
      - subscriptions
swagger: "2.0"